Convert logging.init_app to use dictConfig
This allows each app to adjust the log configuration if needed. It moves
us from an imperative log configuration flow to a declarative one - we
just say what we want in a readable dict, and then logging sets it up as
specified. It should be easier to debug logging issues and understand
exactly what loggers we're setting up.

This will initially be used by the data-store to grab the default
logging config and inject it into our new celery worker, as celery's
logger needs to be set up separately (ie not via the flask app).