FS-4621: add application assignment notification types
This template type will be used by the assessment frontend/ store and
the notification service. The story specifies seperate templates for
assigning/ unassigning a assessor user from an application.

Note that needing to update shared utils and a separate flask app in
order to send a different type of email feels fairly tightly coupled --
this process and structure may be better represented as a small util
that can be directly used by the apps needing ot send emails.