Unpin dependencies
FSD-utils is a library package meant to be installed in a variety of
other applications.

Those applications may themselves require other/specific versions of
dependencies. This library should be compatible with as many variations
of dependencies as possible so that it can be installed in all of our
apps with minimal risk of compatibility issues.

So we should not be pinning specific versions here, and instead just say
what range of dep versions should be compatible.