Override the central rangeStrategy=pin
For libraries, we generally want to support ranges of versions rather
than just specific versions of things (the latter makes more sense for
applications).

This change will make renovate support ranges in pyproject.toml, and
replace the range only when a new version of a dependency falls outside
of that range.

Eg if we pinned `Flask~=2.0`, then when Flask 3.0.0 is released it would
bump it to `Flask~=3.0`.