Allow healthcheck to work in host_matching mode
When a Flask app is running in `host_matching` mode, it will only route
inbound requests to registered endpoints that set a `host` parameter on
the route.

In order to register a view against all hosts, you can use route
variables like normal, so `<host>` will match any route and pass the
value through as the variable `host` to the view.