Remove locale selector/translation support
This code all lives in one repo now (pre-award-frontend - ignoring
post-award which is a frontend app but has no translation support).
Because of this, let's just have the code live there rather than in a
utils lib.

This is moving now because we also need to add `host_matching` support
to the language selection logic, and don't need to do it in a separate
lib that makes it harder to pull changes in.