Add platform as supported app temporarily
We want to add the ability to login via SSO to the platform, but we
don't currently have proper domains which will make integrating with
Microsoft AD infeasible.

As a temporary step, we're going to proxy auth via the existing
authenticator service. Once we've got our domains set up, we'll tear
down this logic and build an SSO flow as part of the funding-service
repo directly.

In order to support proxying auth via authenticator, we'll set up the
new funding-service as a safe return app.

We expect this proxying behaviour to last 1-2 weeks.

https://mhclgdigital.atlassian.net/browse/FSPT-393